% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{complement}
\alias{complement}
\title{complement}
\usage{
complement(hex, typeVal = "")
}
\arguments{
\item{hex}{The base color specified as hex}

\item{typeVal}{Can be specified as split or double. Default is empty.}
}
\description{
This function returns the complement color of a rgb color
}
\examples{
complement("#121314")
}
\keyword{complement}

