% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{complementColors}
\alias{complementColors}
\title{complement colors}
\usage{
complementColors(baseColor, count)
}
\arguments{
\item{baseColor}{The base color specified as hex}

\item{count}{Number of colors the palette should contain}
}
\description{
This function generates a color plate with the complement color
}
\examples{
complementColors("#121314", 5)
}
\keyword{color}
\keyword{complement}

