% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{generateMonochromaticColors}
\alias{generateMonochromaticColors}
\title{Monochromatic}
\usage{
generateMonochromaticColors(baseColor, count)
}
\arguments{
\item{baseColor}{The base color specified as hex}

\item{count}{Number of colors the palette should contain}
}
\description{
This function generates a specified number of monochromatic colors for a given base color
}
\examples{
generateMonochromaticColors("#121314", 5)
}
\keyword{color}
\keyword{monochromatic}

