% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{hsv2rgb}
\alias{hsv2rgb}
\title{hsv2rgb convert}
\usage{
hsv2rgb(h, s, v)
}
\arguments{
\item{h}{Hue of the color}

\item{s}{Saturation of the color}

\item{v}{Value of the color}
}
\description{
This function converts the values of a color from hsv color space to rgb.
}
\examples{
hsv2rgb(150,0.2,0.6)
}
\keyword{hsv}
\keyword{rgb}

