% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{pentadic}
\alias{pentadic}
\title{pentadic}
\usage{
pentadic(hex)
}
\arguments{
\item{hex}{The base color specified as hex}
}
\description{
This function returns pentadic colors to a given hex color
}
\examples{
pentadic("#121314")
}
\keyword{pentadic}

