% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{tetradic}
\alias{tetradic}
\title{tetradic}
\usage{
tetradic(hex)
}
\arguments{
\item{hex}{The base color specified as hex}
}
\description{
This function returns tetradic colors to a given hex color
}
\examples{
tetradic("#121314")
}
\keyword{tetradic}

