% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{triadic}
\alias{triadic}
\title{triadic}
\usage{
triadic(hex)
}
\arguments{
\item{hex}{The base color specified as hex}
}
\description{
This function returns triadic colors to a given hex color
}
\examples{
triadic("#121314")
}
\keyword{triadic}

