% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/palatte-package.R
\name{triadicColors}
\alias{triadicColors}
\title{triadic colors}
\usage{
triadicColors(baseColor, count)
}
\arguments{
\item{baseColor}{The base color specified as hex}

\item{count}{Number of colors the palette should contain}
}
\description{
This function generates a specified number of triadic colors for a given base color
}
\examples{
triadicColors("#121314", 5)
}
\keyword{color}
\keyword{triadic}

